// src/payload-types.ts
export interface ClientModel{
  Id: number;
  UserName: string;
  Email: string;
  FullName: string;
  Phone: string; 
  NationalityId: string;
  HealthConcern: string;
  Age: string;
  Gender: string;
  HeartRate: number; 
  BloodPressure: string;
  Temperature: number;  
  OxygonSaturation: string;  
  ReportedSymptoms: string;   
}